Added main hero
