+++ src/App.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import "./App.css";
-import { MainHero } from "./components/MainHero";
+import { MainHero } from "./components/MainHero/MainHero";
 
 function App() {
   return (
