--- src/components/MainHero.css
@@ -1,81 +1,0 @@
-.main-hero section .container {
-  display: flex;
-  transition: all 0.3s ease-in-out;
-  justify-content: center;
-  flex-direction: column;
-}
-
-@media screen and (min-width: 768px) {
-  .main-hero section .container {
-    height: 100vh;
-  }
-}
-
-@media screen and (min-width: 1024px) {
-  .main-hero section .container {
-    height: 80vh;
-  }
-}
-
-.main-hero section .container .loaders-wrapper {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.loaders-wrapper .loader-text h1 span,
-.is-Architects-daughter-container p span {
-  transition: all 0.3s ease-in-out;
-  animation-timing-function: linear;
-  animation-duration: 1s;
-  animation-iteration-count: 1;
-  display: inline-block;
-  transform: translateY(-10px);
-  opacity: 0;
-}
-
-@media screen and (min-width: 1024px) {
-  .loader-text h1 span:hover,
-  .loader-text h1 span:focus-visible,
-  .is-Architects-daughter-container p span:hover,
-  .is-Architects-daughter-container p span:focus-visible {
-    animation-name: bounce;
-    cursor: pointer;
-    transition: all 0.3s ease-in-out;
-    color: var(--color-sand);
-  }
-}
-
-.is-Architects-daughter {
-  position: relative;
-  z-index: 1;
-}
-
-@media screen and (min-width: 1024px) {
-  .is-Architects-daughter {
-    padding-top: 3rem;
-  }
-}
-
-@media screen and (min-width: 1024px) {
-  svg{
-    width: 100%;
-  }
-}
-
-.is-Architects-daughter-line {
-  position: relative;
-}
-
-.is-Architects-daughter .is-Architects-daughter-line svg {
-  position: absolute;
-  height: 2px;
-  background-color: var(--color-red);
-  color: var(--color-red);
-  bottom: 0;
-  left: 0;
-  z-index: 1;
-  width: 20%;
-  height: auto;
-}
