+++ src/components/MainHero/DChart.jsx
@@ -1,0 +1,25 @@
+import React from "react";
+
+const DChart = () => {
+  return (
+    <svg
+      width='250'
+      height='auto'
+      viewBox='0 0 189 250'
+      fill='none'
+      xmlns='http://www.w3.org/2000/svg'
+    >
+      <path
+        id='d-fill'
+        d='M178.639 105.869C178.639 129.99 174 150.156 164.723 166.367C155.543 182.48 142.213 194.639 124.732 202.842C107.252 210.947 86.207 215 61.5977 215H0.953125V0.839844H68.1895C90.6504 0.839844 110.133 4.84375 126.637 12.8516C143.141 20.7617 155.934 32.5293 165.016 48.1543C174.098 63.6816 178.639 82.9199 178.639 105.869ZM131.471 107.041C131.471 91.2207 129.127 78.2324 124.439 68.0762C119.85 57.8223 113.014 50.2539 103.932 45.3711C94.9473 40.4883 83.8145 38.0469 70.5332 38.0469H46.3633V177.5H65.8457C88.0137 177.5 104.469 171.592 115.211 159.775C126.051 147.959 131.471 130.381 131.471 107.041Z'
+      />
+      <path
+        id='d-line'
+        d='M164.723 166.367L162.987 165.374L162.985 165.377L164.723 166.367ZM124.732 202.842L125.574 204.656L125.582 204.652L124.732 202.842ZM0.953125 215H-1.04688V217H0.953125V215ZM0.953125 0.839844V-1.16016H-1.04688V0.839844H0.953125ZM126.637 12.8516L125.764 14.651L125.772 14.6551L126.637 12.8516ZM165.016 48.1543L163.286 49.1594L163.289 49.1641L165.016 48.1543ZM124.439 68.0762L122.614 68.8933L122.619 68.9038L122.624 68.9143L124.439 68.0762ZM103.932 45.3711L102.977 47.1284L102.985 47.1326L103.932 45.3711ZM46.3633 38.0469V36.0469H44.3633V38.0469H46.3633ZM46.3633 177.5H44.3633V179.5H46.3633V177.5ZM115.211 159.775L113.737 158.423L113.731 158.43L115.211 159.775ZM176.639 105.869C176.639 129.741 172.048 149.541 162.987 165.374L166.458 167.361C175.952 150.771 180.639 130.24 180.639 105.869H176.639ZM162.985 165.377C154.018 181.116 141.006 192.996 123.883 201.031L125.582 204.652C143.42 196.282 157.068 183.844 166.46 167.357L162.985 165.377ZM123.891 201.027C106.737 208.982 85.9942 213 61.5977 213V217C86.4199 217 107.767 212.913 125.574 204.656L123.891 201.027ZM61.5977 213H0.953125V217H61.5977V213ZM2.95312 215V0.839844H-1.04688V215H2.95312ZM0.953125 2.83984H68.1895V-1.16016H0.953125V2.83984ZM68.1895 2.83984C90.4146 2.83984 109.586 6.80144 125.764 14.6509L127.51 11.0522C110.68 2.88606 90.8862 -1.16016 68.1895 -1.16016V2.83984ZM125.772 14.6551C141.919 22.3942 154.409 33.8859 163.287 49.1593L166.745 47.1492C157.458 31.1727 144.362 19.1292 127.501 11.048L125.772 14.6551ZM163.289 49.1641C172.149 64.3115 176.639 83.1787 176.639 105.869H180.639C180.639 82.6612 176.046 63.0518 166.742 47.1445L163.289 49.1641ZM133.471 107.041C133.471 91.043 131.104 77.7424 126.255 67.2381L122.624 68.9143C127.15 78.7224 129.471 91.3984 129.471 107.041H133.471ZM126.265 67.2591C121.512 56.6403 114.385 48.7206 104.879 43.6095L102.985 47.1326C111.642 51.7872 118.187 59.0042 122.614 68.8933L126.265 67.2591ZM104.887 43.6138C95.5345 38.5312 84.0513 36.0469 70.5332 36.0469V40.0469C83.5776 40.0469 94.36 42.4454 102.977 47.1283L104.887 43.6138ZM70.5332 36.0469H46.3633V40.0469H70.5332V36.0469ZM44.3633 38.0469V177.5H48.3633V38.0469H44.3633ZM46.3633 179.5H65.8457V175.5H46.3633V179.5ZM65.8457 179.5C88.3511 179.5 105.44 173.496 116.691 161.121L113.731 158.43C103.497 169.687 87.6763 175.5 65.8457 175.5V179.5ZM116.685 161.127C127.989 148.804 133.471 130.656 133.471 107.041H129.471C129.471 130.105 124.112 147.114 113.737 158.423L116.685 161.127Z'
+        fill='#F0371A'
+      />
+    </svg>
+  );
+};
+
+export default DChart;
