--- src/css/cta.css
@@ -1,15 +1,0 @@
-.project-link {
-  background-color: var(--color-red);
-  border-radius: 2.5rem;
-  padding: 1rem 1.5rem;
-  color: var(--color-white);
-  font-family: var(--font-san-serif);
-  font-weight: 600;
-  display: inline-flex;
-  width: max-content;
-  text-decoration: none;
-}
-
-.project-link:hover {
-  background-color: darken(var(--color-red), 10%);
-}
