--- src/css/global.css
@@ -1,3 +1,0 @@
-@import './root.css';
-@import './typography.css';
-@import './cta.css';
