--- src/css/typography.css
@@ -1,50 +1,0 @@
-p, a, span, ul li, blockquote {
-  font-weight: 400;
-  font-family: var(--font-san-serif);
-  font-size: clamp(1rem, 2vw, 2.2rem);
-  line-height: clamp(1.5rem, 3vw, 3rem);
-}
-
-p.is-bigger {
-  font-size: clamp(3.5rem, 7vw, 8rem);
-  line-height: clamp(4rem, 8vw, 10rem);
-}
-
-h2, h3 {
-  font-family: var(--font-cursive-decorator);
-  margin: 0;
-}
-
-h2 {
-  font-size: clamp(2.8rem, 3vw, 3.5rem);
-  line-height: clamp(3.2rem, 4vw, 5.5rem);
-}
-
-h3 {
-  font-size: clamp(2.6rem, 3vw, 2.8rem);
-  line-height: clamp(3.2rem, 4vw, 5.5rem);
-}
-
-.is-bigger-sans {
-  font-family: var(--font-san-serif);
-  font-size: clamp(2.4rem, 4.5vw, 6.5rem);
-  line-height: clamp(4rem, 6vw, 6rem);
-}
-
-.is-bigger-cursive {
-  font-family: var(--font-cursive);
-  font-size: clamp(2.4rem, 4.5vw, 6.5rem);
-  line-height: clamp(1rem, 1vw, 1rem);
-}
-
-.is-cursive-code {
-  font-family: var(--font-cursive);
-}
-
-.is-block {
-  display: block;
-}
-
-.is-ff-red {
-  color: var(--color-red);
-}
