+++ src/scss/base/containers.scss
@@ -1,0 +1,7 @@
+.container {
+    max-width: $max-width;
+    margin: 0 2rem;
+    @media screen and (min-width: $desktop-extra-big-BP){
+        margin: 0 auto;
+    }
+}
