+++ src/scss/base/typography.scss
@@ -1,0 +1,169 @@
+/*-----------------------------------*\
+    Google Fonts Typography Links
+\*-----------------------------------*/
+
+//* Open Sans //*
+//* font-family: 'Public Sans', sans-serif; //*
+@import url('https://fonts.googleapis.com/css2?family=Public+Sans:ital,wght@0,400;0,600;0,700;1,400;1,600;1,700&display=swap');
+
+//* La Belle Aurore *//
+//* font-family: 'La Belle Aurore', cursive; //*
+@import url('https://fonts.googleapis.com/css2?family=La+Belle+Aurore&display=swap');
+
+//* Tilt Warp *//
+//* font-family: 'Tilt Warp', cursive;
+@import url('https://fonts.googleapis.com/css2?family=Tilt+Warp&display=swap');
+
+//*font-family: 'Fjalla One', sans-serif;
+
+@import url('https://fonts.googleapis.com/css2?family=Fjalla+One&display=swap');
+
+
+p, a, span, ul li, blockquote {
+    margin: 0;
+    font-weight: 400;
+    font-family: var(--font-san-serif);
+    @include fontsize(2rem, 2rem, 2.2rem);
+    @include lineheight(2.4rem, 2.6rem, 3rem);
+
+    &.nav-items {
+        @include fontsize(1.8rem, 1.8rem, 2.2rem);
+        @include lineheight(2.4rem, 2.4rem, 2.5rem);
+    }
+}
+
+p {
+    padding-bottom: 1.5rem;
+    &:last-child {
+        padding-bottom: 0;
+    }
+    &.is-bigger {
+        @include fontsize(3.5rem, 4.5rem, 8rem);
+        @include lineheight(4rem, 4rem, 10rem);
+    }
+}
+
+blockquote {
+    font-family: var(--font-quote);
+    text-indent: 0rem;
+    padding-top: 5rem;
+
+    @media screen and (min-width: $desktop-BP){
+        max-width: 35%;
+        text-indent: -0.9rem;
+    }
+
+    & > p {
+        font-family: var(--font-quote);
+        border-left: 3px solid var(--color-red);
+        padding-left: 1.5rem;
+        
+        @media screen and (min-width: $desktop-BP){
+            padding-left: 3rem;
+        }
+    }
+}
+
+
+ul {
+    margin: 0;
+    padding: 0;
+    li {
+        list-style-type: none;
+    }
+}
+
+a {
+    text-decoration: none;
+
+    &.header-links {
+        @include fontsize(1.8rem, 1.8rem, 2.5rem);
+    }
+}
+
+h2 {
+    @include fontsize(2.8rem, 3rem, 5rem);
+    @include lineheight(3.2rem, 3.5rem, 5.5rem);
+    font-family: var(--font-cursive-decorator);
+    margin: 0;
+}
+
+h3 {
+    @include fontsize(2.6rem, 2.6rem, 4rem);
+    @include lineheight(3.2rem, 3.2rem, 5.5rem);
+    font-family: var(--font-cursive-decorator);
+    margin: 0;
+}
+
+h4 {
+    @include fontsize(2.3rem, 2.3rem, 3rem);
+    @include lineheight(3rem, 3rem, 4.5rem);
+    font-family: var(--font-cursive-decorator);
+    margin: 0;
+}
+
+h5 {
+    @include fontsize(2rem, 2rem, 2.5rem);;
+    @include lineheight(2rem, 2rem, 3.5rem);
+    font-family: var(--font-cursive-decorator);
+    margin: 0;
+}
+
+h6 {
+    @include fontsize(1.8rem, 1.8rem, 2rem);
+    @include lineheight(2.5rem, 2.5rem, 3.5rem);
+    font-family: var(--font-cursive-decorator);
+    margin: 0;
+}
+
+.is-bigger {
+    @include fontsize(2.4rem, 4.5rem, 6.5rem);
+    @include lineheight(4rem, 6rem, 8rem);
+    font-family: 'Tilt Warp', cursive;
+}
+
+.is-color-red {
+    color: $color-red;
+}
+
+.is-Architects-daughter {
+    font-family: var(--font-cursive);
+}
+
+.is-dk-right-3 {
+    padding-right: 1rem;
+    @media screen and (min-width: $desktop-BP){
+        padding-right: 3rem;
+    }
+}
+
+.is-dk-left-3 {
+    padding-left: 1rem;
+    @media screen and (min-width: $desktop-BP){
+        padding-left: 3rem;
+    }
+}
+
+.is-bigger-sans {
+    font-family: var(--font-san-serif);
+    font-size: clamp(2.4rem, 4.5vw, 4.5rem);
+    line-height: clamp(4rem, 6vw, 5rem);
+}
+
+.is-bigger-cursive {
+font-family: var(--font-cursive);
+font-size: clamp(2.4rem, 4.5vw, 6.5rem);
+line-height: clamp(1rem, 1vw, 1rem);
+}
+
+.is-cursive-code {
+    font-family: var(--font-cursive);
+}
+
+.is-block {
+    display: block;
+}
+
+.is-ff-red {
+    color: var(--color-red);
+}
