+++ src/scss/helpers/animations.scss
@@ -1,0 +1,22 @@
+@keyframes bounce {
+    20%,
+    40%,
+    50%,
+    70%,
+    100%,
+    to {
+        transform: scale(1, 1);
+    }
+    40% {
+        transform: scale(1.75, 0.65);
+    }
+    45% {
+        transform: scale(1.75, 0.65);
+    }
+    70% {
+        transform: scale(1.25, 0.75);
+    }
+    90% {
+        transform: scale(1.15, 0.85);
+    }
+}
