+++ src/scss/helpers/btns-and-hovers.scss
@@ -1,0 +1,45 @@
+.project-link {
+    @include fontsize(1.6rem, 1.8rem, 1.8rem);
+    background-color: var(--color-red);
+    font-family: var(--font-san-serif);
+    border-radius: 2.5rem;
+    padding: 1rem 1.5rem;
+    color: var(--color-white);
+    font-weight: 600;
+    display: inline-flex;
+    width: max-content;
+}
+
+.hvr-underline-from-left {
+    text-decoration: none;
+    display: inline-block;
+    vertical-align: middle;
+    transform: perspective(1px) translateZ(0);
+    box-shadow: 0 0 1px rgba(0, 0, 0, 0);
+    position: relative;
+    overflow: hidden;
+    transition: all 300ms ease-in-out;
+    
+    &:before {
+        content: "";
+        position: absolute;
+        z-index: -1;
+        left: 0;
+        right: 100%;
+        bottom: 0;
+        background: var(--color-red);
+        height: 3px;
+        transition-property: right;
+        transition-duration: 0.3s;
+        transition: all 300ms ease-in-out;
+    }
+
+    &:hover, &:focus, &:focus-visible, &:active {
+        color: var(--color-red);
+        transform: scale(0.9);
+        transition: all 300ms ease-in-out;
+        &::before {
+            right: 0;
+        }
+    }
+}
