+++ src/scss/variables.scss
@@ -1,0 +1,6 @@
+$color-red: var(--color-red);
+$color-white: var(--color-white);
+
+/// Container's maximum width
+/// @type Length
+$max-width: 1380px;
